Convert tasks to a set
Instead of `tasks` being a map from TaskId to Task, it is now a set of
tasks, each of which has an `id' field.

This makes the spec a bit easier to read, and might allow some
optimisations in the model checking later.

Also, group transitions by component. Check that every state is handled
by only one component, and that no component tries to perform a
transition owned by another.

Signed-off-by: Thomas Leonard <thomas.leonard@docker.com>