Improve TLA+ model documentation and performance
Extend the model checking section to show more ways of improving
performance: disabling actions, sharing task IDs and combining states.

Move each component into a sub-module, so that it is clear they are
independent. Move some helper operators that were in the orchestrator
section (but also used elsewhere) to the types section.

Signed-off-by: Thomas Leonard <thomas.leonard@docker.com>